Fix recursion limits, Return propagation, and numeric edge cases (#718)
* Fix six divergences found by running Project Euler solutions

Running Nayuki's 126 Wolfram Language Project Euler solutions
(github.com/nayuki/Project-Euler-solutions) through Woxi surfaced six
independent engine bugs. Each fix comes with a regression test.

- Multiple `Sum` / `Product` iterators are enumerated outermost-first.
  Both are `HoldAll`, so Wolfram substitutes the outer iterator's value
  into the still-held inner sum before evaluating it. Evaluating the
  innermost iterator first left the outer variable symbolic, which let a
  catch-all definition (`f[0, _, _] = 0`) win over the specific one
  (`f[0, 0, 0] = 1`) and silently returned a wrong total. Iterators that
  are symbolic, or whose grid is too large to enumerate, keep using the
  symbolic inside-out route.

- `$RecursionLimit` is honoured instead of being a hard-wired 1024, and
  reads as Wolfram's default of 1024 rather than 256. `Infinity` removes
  the limit; the environment is only consulted once a recursion passes
  depth 20, the smallest limit Wolfram accepts.

- `Floor` / `Ceiling` of an exact algebraic number that sits a hair from
  an integer are decided at arbitrary precision. `(10^15 - 1)^(1/15)`
  rounds to exactly `10.0` in double precision though it is strictly
  below 10, so `Floor` answered 10. The precision now grows until the two
  sides of the integer separate.

- `FromContinuedFraction` accumulates its convergents in `BigInt`. The
  numerator of `ContinuedFraction[E, 100]` outgrows i128 long before the
  last term, and the multiply panicked in a build with overflow checks.

- `ContinuedFraction[x, n]` of a quadratic irrational comes from the
  exact periodic recurrence. `ContinuedFraction[Sqrt[2], 200]` is 200
  exact 2s; the f64 fallback drifted off the true expansion after ~20
  terms.

- `Return` keeps travelling up through `;` and the loops. `For` and
  `While` turn the signal into a literal `Return[val]`, but the
  `CompoundExpr` (`;`) path and the loops themselves dropped it, so a
  `Return` out of a nested `For` was swallowed and iteration continued.
  `Do` and `Scan` still consume it, as Wolfram specifies.

- `Range` no longer refuses lists past a million elements. Wolfram bounds
  it only by memory, and `Range[10^7]` is ordinary in this kind of code.

* Keep rational-exponent extraction exact for large exponents

`n^(p/q)` pulls the whole part of the exponent out in front, but the
extraction cast the numerator to u64 and accumulated the extracted factor
in i128:

- A numerator past `u64::MAX` wrapped, so
  `2^(29904177082838066039/10^18)` simplified as if the exponent were
  11.457… instead of 29.904…, giving 2812.1 where the true value is
  1.0047*10^9. Project Euler 197 iterates exactly this power and landed
  on the wrong fixed point.
- An extracted part beyond i128 panicked outright: `2^(1000/7)` pulls out
  2^142 and aborted with "attempt to multiply with overflow".

The exponent arithmetic is i128 with checked multiplication and the
extracted factor is a `BigInt`, so `2^(1000/7)` now gives Wolfram's
`5575186299632655785383929568162090376495104*2^(6/7)`. An extracted
exponent too large to write down leaves the power symbolic.

Trial division of the base is bounded at 10^6: `(10^19 - 1)^(1/19)` is
3^2 * 1111111111111111111, and grinding the second (prime) factor out
took ~3*10^9 steps for a result that is identical either way, since a
cofactor kept whole splits into the same extracted and radical parts.
`Floor[(10^19 - 1)^(1/19)]` drops from 20 s to 0.08 s.

+++ src/evaluator/listable.rs
@@ -493,7 +493,7 @@ pub fn is_system_variable_name(name: &str) -> bool {
 /// Look up system $ variables
 pub fn get_system_variable(name: &str) -> Option<Expr> {
   match name {
-    "$RecursionLimit" => Some(Expr::Integer(256)),
+    "$RecursionLimit" => Some(Expr::Integer(1024)),
     "$IterationLimit" => Some(Expr::Integer(4096)),
     "$HistoryLength" => Some(Expr::Identifier("Infinity".to_string())),
     // Wolframscript runs each script as a fresh session, so `$Line` —
